% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_index.R
\name{cor_index}
\alias{cor_index}
\title{Corrected indices}
\usage{
cor_index(host, nonhost, format = "wide", scale = TRUE,
  years_as_rownames = FALSE)
}
\arguments{
\item{host}{df: ring-width indices}

\item{nonhost}{df: master chronology}

\item{format}{c('wide', 'long'): input data format}

\item{scale}{logical: flag to normalize}

\item{years_as_rownames}{logical: flag for dplR format}
}
\description{
Use tree-level host ring-width indices and a single master
nonhost chronology to create corrected indices for each
host tree.
}
\examples{
\dontrun{
cor_index(ob_host, ob_nonhost)
cor_index(ob_host \%>\%
            gather('id', 'rwi', -Year),
          ob_nonhost,
          format = 'long')
}
}
