% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ob_host}
\alias{ob_host}
\title{Ring-width indices from four Douglas-fir sites in the Okanogan Highlands, WA, USA}
\format{Dataframe with columns for each sampled tree.
\describe{
 \item{year}{A value between 1685 and 2014.}
 \item{MPD01}{Time series for tree 01 sampled at site MPD. Format repeats for each tree across four sites.}
}}
\source{
T. Michael Ellis -- collected Aug. - Oct. 2014 \url{http://toddmellis.wordpress.com}
}
\usage{
ob_host
}
\description{
Ring-width indices from four Douglas-fir sites in the Okanogan Highlands, WA, USA
}
\examples{
\dontrun{
 ob_host
}
}
\keyword{datasets}
