% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ob_nonhost}
\alias{ob_nonhost}
\title{First principal component forming a master chronology from four ponderosa pine sites in the Okanogan Highlands, WA, USA}
\format{Dataframe with year and master chronology columns.
\describe{
 \item{year}{A value between 1685 and 2014.}
 \item{nonhost}{Time series for first principal component of a master nonhost ponderosa pine chronology.}
}}
\source{
T. Michael Ellis -- collected Aug. - Oct. 2014 \url{http://toddmellis.wordpress.com}
}
\usage{
ob_nonhost
}
\description{
First principal component forming a master chronology from four ponderosa pine sites in the Okanogan Highlands, WA, USA
}
\examples{
\dontrun{
 ob_nonhost
}
}
\keyword{datasets}
