% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{outbreak}
\alias{outbreak}
\title{Outbreak reconstruction tools}
\usage{
outbreak(ci, min = 4, sd = -1.28, prop = TRUE)
}
\arguments{
\item{ci}{df: corrected indices}

\item{min}{int: minimum length of outbreaks}

\item{sd}{dbl: minimum standard deviation}

\item{prop}{logical: controls output format}
}
\description{
Using corrected indices produced by cor_index, this
function produces defoliating insect outbreak
reconstructions at the tree level.
}
\examples{
\dontrun{
outbreak( cor_index( ob_host, ob_nonhost))

foo <- cor_index( ob_host, ob_nonhost))
sites <- foo \%>\%
  mutate( site = substr( tree_id, 1, 3)) \%>\%
  .$site \%>\%
  unique()
bar <- list()
for (i in 1:length(sites)) {
  bar[[i]] <- foo \%>\%
    mutate(site = substr(tree_id, 1, 3)) \%>\%
    filter(site == groups[i]) \%>\%
    outbreak() \%>\%
    mutate(site = groups[i])
}
outbreaks <- bind_rows(bar)
}
}
